rpm/backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
faiss-cpu==1.11.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
